# internal/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Broto API",
    "version": "1.0.0",
    "description": "Contrato HTTP do backend Go. Sessões opacas via Bearer, não JWT Supabase. Try it out executa operações reais, incluindo exclusões e IA. Integrações dependem da configuração do servidor. Veja docs/api.md e docs/google-auth.md."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesmo servidor da documentação"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/auth/logout": {
      "post": {
        "operationId": "post_v1_auth_logout",
        "summary": "Encerrar sessão",
        "tags": [
          "Autenticação"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "post_v1_auth_refresh",
        "summary": "Rotacionar sessão",
        "tags": [
          "Autenticação"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/auth/user": {
      "get": {
        "operationId": "get_v1_auth_user",
        "summary": "Ler usuário da sessão",
        "tags": [
          "Autenticação"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/auth/password": {
      "patch": {
        "operationId": "patch_v1_auth_password",
        "summary": "Alterar senha e invalidar sessões anteriores",
        "tags": [
          "Autenticação"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/google/start": {
      "post": {
        "operationId": "post_v1_auth_google_start",
        "summary": "Iniciar Google OAuth com PKCE",
        "tags": [
          "Google OAuth"
        ],
        "description": "PKCE S256. Abra a URL retornada no navegador para receber o cookie. Retorno deve estar em GOOGLE_RETURN_URLS. Ver docs/google-auth.md; 503 quando desativado.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code_challenge": {
                    "type": "string",
                    "minLength": 43,
                    "maxLength": 43
                  },
                  "redirect_uri": {
                    "type": "string"
                  }
                },
                "required": [
                  "code_challenge",
                  "redirect_uri"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/auth/google/authorize": {
      "get": {
        "operationId": "get_v1_auth_google_authorize",
        "summary": "Redirecionamento OAuth: authorize",
        "tags": [
          "Google OAuth"
        ],
        "description": "Fluxo de navegação com cookie HttpOnly e estado de uso único. Abra no navegador; não use Try it out para completar login Google.",
        "responses": {
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "302": {
            "description": "Redirecionamento ao Google ou ao retorno permitido do app",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL de destino"
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      }
    },
    "/v1/auth/google/callback": {
      "get": {
        "operationId": "get_v1_auth_google_callback",
        "summary": "Redirecionamento OAuth: callback",
        "tags": [
          "Google OAuth"
        ],
        "description": "Fluxo de navegação com cookie HttpOnly e estado de uso único. Abra no navegador; não use Try it out para completar login Google.",
        "responses": {
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "302": {
            "description": "Redirecionamento ao Google ou ao retorno permitido do app",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL de destino"
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": ""
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      }
    },
    "/v1/auth/google/exchange": {
      "post": {
        "operationId": "post_v1_auth_google_exchange",
        "summary": "Trocar código temporário por sessão Broto",
        "tags": [
          "Google OAuth"
        ],
        "description": "Código de uso único expira em 60 segundos; verifier original do app.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "code_verifier": {
                    "type": "string",
                    "minLength": 43,
                    "maxLength": 128
                  }
                },
                "required": [
                  "code",
                  "code_verifier"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/account": {
      "delete": {
        "operationId": "delete_v1_account",
        "summary": "Excluir conta e seus dados",
        "tags": [
          "Conta"
        ],
        "description": "Operação destrutiva. Fotos são removidas por fila persistente; 202 indica limpeza pendente.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "202": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "photos_cleanup_pending": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/functions/v1/delete-account": {
      "post": {
        "operationId": "post_functions_v1_delete_account",
        "summary": "Excluir conta e seus dados",
        "tags": [
          "Conta"
        ],
        "description": "Operação destrutiva. Fotos são removidas por fila persistente; 202 indica limpeza pendente.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "202": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "photos_cleanup_pending": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/species-facts": {
      "get": {
        "operationId": "get_v1_species_facts",
        "summary": "Consultar ficha de espécie em cache",
        "tags": [
          "Espécies"
        ],
        "description": "Retorna a ficha diretamente ou null; não inicia uma análise paga.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cuidados": {
                      "type": "object",
                      "properties": {
                        "rega_dias": {
                          "type": "integer",
                          "description": "intervalo entre regas, de 1 a 60 dias"
                        },
                        "luz": {
                          "type": "string",
                          "enum": [
                            "sol_direto",
                            "sol_parcial",
                            "luz_indireta",
                            "meia_sombra",
                            "sombra"
                          ]
                        },
                        "luz_nota": {
                          "type": "string",
                          "description": "uma frase completa com ponto final, ou vazio"
                        },
                        "adubo": {
                          "type": "string",
                          "enum": [
                            "quinzenal",
                            "mensal",
                            "bimestral",
                            "estacional",
                            "nao_precisa"
                          ]
                        },
                        "adubo_nota": {
                          "type": "string",
                          "description": "uma frase completa com ponto final, ou vazio"
                        },
                        "vaporizar_dias": {
                          "type": "integer",
                          "description": "intervalo entre vaporizações, de 1 a 15 dias; null quando a espécie não precisa",
                          "nullable": true
                        },
                        "girar_dias": {
                          "type": "integer",
                          "description": "intervalo para girar o vaso, de 7 a 60 dias"
                        },
                        "replantar_meses": {
                          "type": "integer",
                          "description": "intervalo para trocar de vaso, de 6 a 48 meses"
                        },
                        "podar_mes": {
                          "type": "integer",
                          "description": "mês do ano em que a poda deve ser feita, de 1 a 12; null quando a espécie não é podada",
                          "nullable": true
                        }
                      },
                      "required": [
                        "rega_dias",
                        "luz",
                        "luz_nota",
                        "adubo",
                        "adubo_nota",
                        "vaporizar_dias",
                        "girar_dias",
                        "replantar_meses",
                        "podar_mes"
                      ],
                      "additionalProperties": false
                    },
                    "toxica_para_pets": {
                      "type": "boolean"
                    },
                    "temperatura": {
                      "type": "object",
                      "properties": {
                        "min_c": {
                          "type": "integer",
                          "description": "menor temperatura confortável, em Celsius"
                        },
                        "max_c": {
                          "type": "integer",
                          "description": "maior temperatura confortável, em Celsius"
                        },
                        "nota": {
                          "type": "string",
                          "description": "uma frase completa com ponto final"
                        }
                      },
                      "required": [
                        "min_c",
                        "max_c",
                        "nota"
                      ],
                      "additionalProperties": false
                    },
                    "cultivo": {
                      "type": "string",
                      "description": "duas a quatro frases sobre como multiplicar e manter"
                    },
                    "simbolismo": {
                      "type": "string",
                      "description": "atribuição cultural declarada como atribuição, ou string vazia"
                    }
                  },
                  "required": [
                    "cuidados",
                    "toxica_para_pets",
                    "temperatura",
                    "cultivo",
                    "simbolismo"
                  ],
                  "additionalProperties": false,
                  "nullable": true
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "scientific",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "description": ""
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "pt-BR",
                "en-US",
                "es-ES"
              ]
            },
            "description": ""
          }
        ]
      }
    },
    "/v1/reminders/unread-count": {
      "get": {
        "operationId": "get_v1_reminders_unread_count",
        "summary": "Contar lembretes não lidos",
        "tags": [
          "Lembretes"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/reminders/read": {
      "post": {
        "operationId": "post_v1_reminders_read",
        "summary": "Marcar todos os lembretes como lidos",
        "tags": [
          "Lembretes"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "updated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/photos": {
      "post": {
        "operationId": "post_v1_photos",
        "summary": "Enviar foto binária",
        "tags": [
          "Fotos"
        ],
        "description": "JPEG, PNG ou WebP, máximo 8 MiB. Envie bytes diretamente, não multipart/form-data. Consome quota de armazenamento.",
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/webp": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_v1_photos",
        "summary": "Excluir foto sem referências",
        "tags": [
          "Fotos"
        ],
        "description": "409 enquanto a foto estiver vinculada a um registro.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OK"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string"
                  }
                },
                "required": [
                  "path"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/photos/sign": {
      "post": {
        "operationId": "post_v1_photos_sign",
        "summary": "Criar URL temporária para foto própria",
        "tags": [
          "Fotos"
        ],
        "description": "CloudFront: 300 segundos. Armazenamento local: 3600 segundos. Grave apenas path, nunca a URL assinada.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "signedUrl": {
                      "type": "string",
                      "format": "uri"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "enum": [
                        300,
                        3600
                      ]
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string"
                  }
                },
                "required": [
                  "path"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cole somente o access_token de uma sessão Broto. Não use chaves de provedores."
      }
    },
    "schemas": {
...
      "OK": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "email_confirmed_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "user_metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "bearer"
            ]
          },
          "expires_in": {
            "type": "integer"
          },
          "expires_at": {
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "expires_at",
          "user"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Erro JSON: 400 entrada inválida; 401 sessão; 402 créditos/plano; 403 acesso/origem; 404 inexistente; 409 conflito; 429 limite; 500 erro interno; 502 provedor de IA; 503 integração/banco/concorrência; 507 quota de fotos. Consulte erro para o código específico.",
        "headers": {
          "Retry-After": {
            "description": "Segundos até nova tentativa quando aplicável",
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "erro": "sem_token"
            }
          }
        }
      }
    }
  }
}